==> 01-installing-all-dependencies.ipynb <==
# Connect to MongoDB - Atlas
Este notebook se encarga de poder conectar a mongoDB - Atlas para poder descargar el dataset que necesitamos para este proyecto.

## Getting your current IP

# !curl ifconfig.io

# !curl ipecho.net/plain

!python -V

 ## Update system apps
 
> If can not input the string, copy this command to the terminal

#!sudo apt-get update

# !sudo apt-get upgrade

---

## Installing libraries

!pip install pymongo

!pip install pandas


!pip install numpy

!pip install html5lib

!pip install nltk
