==> 02-connect-to-MongoDB-download-reviews.ipynb <==
# 2. Connect to MongoDB
The purporse of this notebook is to download the dataset from our private MongoDB-Atlas to the local computer, then to convert it to a .csv file.

import pymongo
import os
import pandas as pd
import numpy as np

file_credential = ".my-mongo-credential"
file_out = "./CSV/df_amazon_original.csv"

f=open(file_credential,"r")
lines=f.readlines()
username=lines[0].strip()
password=lines[1].strip()
f.close()


# !curl ipconfig.io

# client = pymongo.MongoClient(f'mongodb+srv://{username}:{password}@cluster0.7cwggsq.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0') # AJ
client = pymongo.MongoClient(f'mongodb+srv://{username}:{password}@cluster0.ibac3d0.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0') # MJ


client

print(client.list_database_names())

db = client.Amazon
collection = db.reviews
print(db)
print(collection)



if os.path.exists(file_out):
    print("File already exist, loading from local storage.")
    amazon_df = pd.read_csv(file_out,index_col=0)
else:
    print("Calling collection from MongoDB and converting it to Dataframe")
    amazon_df = pd.DataFrame(list(collection.find()))
    print("Dataframe load complete.")
    print("Trying to save to .csv file.")
    amazon_df.to_csv(file_out)
    

amazon_df.head()

amazon_df.to_csv(file_out, index=False)
