==> 03-data-cleaning.ipynb <==
# Data Preparation


## Data Load

import pandas as pd
import numpy as np

file_in = "./CSV/df_amazon_original.csv"
file_out = "./CSV/df_data-cleaning.csv"

df = pd.read_csv(file_in)
# df.drop("Unnamed: 0",axis=1,inplace=True)

df.head()

## Processing

df.replace({None: np.nan}, inplace=True)
# df = df.set_index("").reset_index(drop=True).drop('_id',axis=1)


df.info()

# Null values
df.isna().sum()

df[df.isna().any(axis=1)]

# Drop Na
df = df.dropna()

# Check for Duplicates
df.duplicated().sum()

## Change data type

df.info()

# Date to datetime
df['Time'] = pd.to_datetime(df['Time'], unit='s')
# Object to String
df['Summary'] = df['Summary'].astype('string')
df['Text'] = df['Text'].astype('string')

df['Year'] = df['Time'].dt.year
df['Month'] = df['Time'].dt.month
df['Day'] = df['Time'].dt.day

df.info()

## Data out

df.to_csv(file_out, index=False)
