==> 04-sentiment-analysis-polarity.ipynb <==
import pandas as pd

file_dataset = "./CSV/df_data-cleaning.csv"
file_out = "./CSV/df_processed_polarity_and_satisfaction.csv"

dataset = pd.read_csv(file_dataset)

dataset.head()

# POLARITY OF THE REVIEWS

#VADER for positive/negative sentiment analysis
import nltk
from nltk.sentiment.vader import SentimentIntensityAnalyzer

nltk.download('vader_lexicon')

#Reviews
reviews = dataset[['Id','Text']]

reviews.head()

#Vader Sentiment Analyzer
analyzer = SentimentIntensityAnalyzer()

#Store the polarity of the reviews
reviews_polarity = []

%%time
#Store the compound polarity for each review
#Note: this takes around 12 minutes (Google Colab)
for index in range(len(reviews)):
  review = reviews.loc[index, 'Text']
  scores = analyzer.polarity_scores(review)
  compound_score = scores['compound']
  reviews_polarity.append(compound_score)

#Check that all reviews have a polarity
print("Calculated polarity of", len(reviews_polarity), "reviews")

#Turn array into dataframe
polarity_dataframe = pd.DataFrame(reviews_polarity, columns = ['Polarity'])

#Add review Ids to the dataframe
polarity_dataframe = pd.concat([reviews["Id"], polarity_dataframe], axis=1)

#Add the polarity to the original dataframe
dataframe_with_polarity = pd.concat([dataset, polarity_dataframe["Polarity"]], axis=1)

polarity_dataframe

polarity_dataframe.head()

# dataframe_with_polarity.drop("Unnamed: 0",axis=1,inplace=True)

dataframe_with_polarity.head()

#Save the polarity onto a csv file
# polarity_dataframe.to_csv('polarity.csv', index=False)  # not needed for now

#Save the whole dataframe with polarity to a csv file
# dataframe_with_polarity.to_csv(file_out, index=False)

## small dataframe resume


polarity_and_score_df = pd.concat([polarity_dataframe, dataset["Score"]], axis=1)

polarity_and_score_df

%%time
customer_satisfied = []

for ind in range(len(polarity_and_score_df)):
  polarity = polarity_and_score_df.loc[ind, 'Polarity']
  score = polarity_and_score_df.loc[ind, 'Score']

  percentage_polarity = ((polarity - (-1)) * 100) / (1 - (-1))
  percentage_score = ((score - (1)) * 100) / (5 - (1))

  average_percentage = (percentage_polarity + percentage_score)/200

  if (average_percentage >= 0.61):
    customer_satisfied.append(1)
  else:
    customer_satisfied.append(0)

# #Create a dataframe with the satisfaction scores
# customer_satisfied = pd.DataFrame(customer_satisfied, columns = ['Customer Satisfied'])

# #Add the review Id
# customer_satisfied = pd.concat([polarity_and_score_df["Id"], satisfied_dataframe], axis=1)

# #Save onto a csv file
# # satisfied_dataframe.to_csv('satisfaction.csv', index=False)

#Create a dataframe with the satisfaction scores
satisfied_dataframe = pd.DataFrame(customer_satisfied, columns = ['Customer_Satisfied'])

#Add the review Id
satisfied_dataframe = pd.concat([polarity_and_score_df["Id"], satisfied_dataframe], axis=1)

#Save onto a csv file
# satisfied_dataframe.to_csv('satisfaction.csv', index=False)

# dataframe_with_polarity.drop("Unnamed: 0",axis=1,inplace=True)

satisfied_dataframe.head()

#Add the customer satisfaction to the original dataframe with polarity
dataframe_with_polarity_and_satisfaction = pd.concat([dataframe_with_polarity, satisfied_dataframe["Customer_Satisfied"]], axis=1)

#Save onto a csv file
# dataframe_with_polarity_and_satisfaction.to_csv('dataframe_with_polarity_and_satisfaction.csv', index=False)

dataframe_with_polarity_and_satisfaction.head()

# #Save onto a csv file
dataframe_with_polarity_and_satisfaction.to_csv(file_out, index=False)

dataframe_with_polarity_and_satisfaction.head()
