==> 05-NLP-word-count.ipynb <==
# COUNT THE OCCURRENCE OF A PARTICULAR WORD IN EACH AND ALL REVIEWS

from sklearn.feature_extraction.text import CountVectorizer

import pandas as pd

file_dataset = "./CSV/df_processed_polarity_and_satisfaction.csv"
file_processed_text = "./CSV/processed_text.csv"
file_vocabulary_processed = "./CSV/vocabulary_processed.csv"
file_term_counts_per_review = "./CSV/term_counts_per_review.csv"
# file_out = "df_processed_sentimentValue.csv"

df = pd.read_csv(file_dataset)
df.head()

# Reviews
reviews = df[['Id', 'Text']]

#Corpus of original reviews
reviews_corpus = reviews['Text'].tolist()
reviews_corpus[0]

## Raw Count

%%time
#Raw word count
vectorizer_raw = CountVectorizer()
raw_counts = vectorizer_raw.fit_transform(reviews_corpus)

feature_names = list(vectorizer_raw.get_feature_names_out())

len(feature_names)

## Count with Preprocessing

### Preprocessing Functions

!pip install html5lib
!pip install lxml
!pip install bs4

import string
import re
import nltk

from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize
from urllib.parse import urlparse
from bs4 import BeautifulSoup

nltk.download('stopwords')
nltk.download('punkt')

stop_words = stopwords.words('english')

#Remove Stop Words
def remove_stopwords(review):
  words = word_tokenize(review)
  words = [word for word in words if word.lower() not in stop_words]
  processed_review = ' '.join(words)

  return processed_review

#Remove Punctuation
def remove_punctuation(review):
  processed_review = review.translate(str.maketrans(string.punctuation, ' '*len(string.punctuation)))
  return processed_review

#Remove numbers
def remove_numbers(review):
  processed_review = ''.join(character for character in review if not character.isdigit())
  return processed_review

#Lowercase sentence
def lowercase_text(text):
    return text.lower()

#Remove Links
def remove_links(review):
  words = review.split()
  for i, word in enumerate(words):
    parsed_url = urlparse(word)
    if parsed_url.scheme and parsed_url.netloc:
      words[i] = ""

  return ' '.join(words)

#Remove html expressions
def remove_html(review):
  process = BeautifulSoup(review, 'html5lib')
  # process = BeautifulSoup(review, 'html.parser')
  processed_review = process.get_text(separator=" ")
  return processed_review

def preprocess_review(review):
  preprocessed_review = remove_html(review)
  # preprocessed_review = remove_links(preprocessed_review)
  preprocessed_review = remove_punctuation(preprocessed_review)
  preprocessed_review = remove_stopwords(preprocessed_review)
  preprocessed_review = remove_numbers(preprocessed_review)
  preprocessed_review = lowercase_text(preprocessed_review)

  return preprocessed_review

### Preprocess the Corpus

%%time
#Preprocess the whole corpus
#Note: 10+ minutes on Colab
preprocessed_corpus = [preprocess_review(review) for review in reviews_corpus]

#Turn array into dataframe
processed_text_dataframe = pd.DataFrame(preprocessed_corpus, columns = ['Processed Text'])

#Add review Ids to the dataframe
processed_text_dataframe = pd.concat([reviews["Id"], processed_text_dataframe], axis=1)

processed_text_dataframe.to_csv(file_processed_text, index=False)

processed_text_dataframe.head(10)

### Count the ocurrences of each term for each and all reviews

%%time
#Processed Word Count
vectorizer_processed = CountVectorizer()
processed_counts = vectorizer_processed.fit_transform(preprocessed_corpus)

#Processed Vocabulary
feature_names_processed = list(vectorizer_processed.get_feature_names_out())

#Save processed vocabulary
vocab_df = pd.DataFrame(feature_names_processed, columns=['Term'])
vocab_df.to_csv(file_vocabulary_processed, index=False)

len(feature_names_processed)

# Saving the Term Count Per Review

import numpy as np

## Saving the counts for all reviews

#This code shoud work if there is enough RAM available
def save_term_counts_csv():
  term_counts_array = processed_counts.toarray()

  #Turn the array into a pandas dataframe
  term_counts_dataframe = pd.DataFrame(term_counts_array, columns=feature_names_processed)

  #Append the review ID at the beginning for cross referencing
  term_counts_dataframe = pd.concat([reviews["Id"], term_counts_dataframe], axis=1)

  #Save the review as
  term_counts_dataframe.to_csv(file_term_counts_per_review, index=False)

## Save the term counts of a particular set of reviews

def get_review_term_counts(review_ids, filename):
  #List the reviews' row Ids
  index_list = [int(df.index[df['Id'] == id][0]) for id in review_ids]

  #Get a list of terms that are in at least one of the reviews
  vocab_id_list = []

  for ind in index_list:
    review_vocab = processed_counts[ind, :]
    vocab_ids = [vocab_id for vocab_id in review_vocab.nonzero()[1]]

    for vocab_id in vocab_ids:
      if vocab_id not in vocab_id_list:
        vocab_id_list.append(vocab_id)

  #For each term, get the count of it in each review
  id_term_counts = {}

  #First put the review Ids
  id_term_counts['Id'] = review_ids

  for vocab_id in vocab_id_list:
    vocab_counts = []

    for ind in index_list:
      vocab_counts.append(processed_counts[int(ind), int(vocab_id)])

    id_term_counts[feature_names_processed[int(vocab_id)]] = vocab_counts

  #Turn into a pandas DataFrame
  counts_dataframe = pd.DataFrame.from_dict(id_term_counts)

  counts_dataframe.to_csv(f'{filename}.csv', index=False)

  return(counts_dataframe)


def get_review_term_counts_raw(review_ids, filename):
  #List the reviews' row Ids
  index_list = [int(df.index[df['Id'] == id][0]) for id in review_ids]

  #Get a list of terms that are in at least one of the reviews
  vocab_id_list = []

  for ind in index_list:
    review_vocab = raw_counts[ind, :]
    vocab_ids = [vocab_id for vocab_id in review_vocab.nonzero()[1]]

    for vocab_id in vocab_ids:
      if vocab_id not in vocab_id_list:
        vocab_id_list.append(vocab_id)

  #For each term, get the count of it in each review
  id_term_counts = {}

  #First put the review Ids
  id_term_counts['Id'] = review_ids

  for vocab_id in vocab_id_list:
    vocab_counts = []

    for ind in index_list:
      vocab_counts.append(raw_counts[int(ind), int(vocab_id)])

    id_term_counts[feature_names[int(vocab_id)]] = vocab_counts

  #Turn into a pandas DataFrame
  counts_dataframe = pd.DataFrame.from_dict(id_term_counts)

  counts_dataframe.to_csv(f'{filename}.csv', index=False)

  return(counts_dataframe)

def find_term(term, filename):
  if term not in feature_names_processed:
    print("The word is not in any reviews")
    return None
  else:
    #Get the term index from the feature names list
    term_index = feature_names_processed.index(term)

    #Reviews where the term exists
    exists_in_reviews = processed_counts[:, term_index]

    review_ids = [review_id for review_id in exists_in_reviews.nonzero()[0]]

    #Get the count per review
    review_counts_dict = {df.loc[rev_id, 'Id']: processed_counts[rev_id, term_index] for rev_id in review_ids}

    counts_dataframe = pd.DataFrame(review_counts_dict, index=[0])
    counts_dataframe['Total'] = counts_dataframe.sum(axis=1, numeric_only=True)
    counts_dataframe.insert(0, "Term", [term], True)

    #Save as csv
    counts_dataframe.to_csv(f'{filename}.csv', index=False)

    return(counts_dataframe)

def find_term_raw(term, filename):

  if term not in feature_names:
    print("The word is not in any reviews")
    return None
  else:
    #Get the term index from the feature names list
    term_index = feature_names.index(term)

    #Reviews where the term exists
    exists_in_reviews = raw_counts[:, term_index]

    review_ids = [review_id for review_id in exists_in_reviews.nonzero()[0]]

    #Get the count per review
    review_counts_dict = {df.loc[rev_id, 'Id']: raw_counts[rev_id, term_index] for rev_id in review_ids}

    counts_dataframe = pd.DataFrame(review_counts_dict, index=[0])
    counts_dataframe['Total'] = counts_dataframe.sum(axis=1, numeric_only=True)
    counts_dataframe.insert(0, "Term", [term], True)

    #Save as csv
    counts_dataframe.to_csv(f'{filename}.csv', index=False)

    return(counts_dataframe)

#Count the terms in Processed Review 69726
get_review_term_counts([69726], 'review_69726')

#Processed Review 69726
review = processed_text_dataframe[processed_text_dataframe['Id'] == 69726]['Processed Text'].iloc[0]
print(review)

#Count the terms in Review 69726
get_review_term_counts_raw([69726], 'review_69726_raw')

#Processed Review 69726
review = reviews[reviews['Id'] == 69726]['Text'].iloc[0]
print(review)

#Find the term 'recommend' in the processed reviews
find_term('recommend', 'recommend')

#Example: processed review 568311 with 2 occurrences of the word 'recommend'
text = processed_text_dataframe.loc[processed_text_dataframe['Id'] == 568311]['Processed Text'].values[0]
print(text)

#Find the term 'recommend' in the reviews
find_term_raw('recommend', 'recommend_raw')

#Example: review 568311 with 2 occurrences of the word 'recommend'
text = df.loc[df['Id'] == 568311]['Text'].values[0]
print(text)
