==> 07-connect-to-MongoDB-upload.ipynb <==
# 6. Connect to MongoDB - UpLoad

The purporse of this notebook is to upload the dataset to our private MongoDB-Atlas from the local computer.

import pymongo
from pymongo import UpdateOne
import os
import pandas as pd
import numpy as np

file_credential = ".my-mongo-credential"
# file_out = "df_amazon_original.csv"

f=open(file_credential,"r")
lines=f.readlines()
username=lines[0].strip()
password=lines[1].strip()
f.close()

client = pymongo.MongoClient(f'mongodb+srv://{username}:{password}@cluster0.ibac3d0.mongodb.net/?retryWrites=true&w=majority&appName=Cluster0') # MJ


client

db = client.Amazon
collection = db.reviews
print(db)
print(collection)

print(client.list_database_names())

## Querry test

# Get one specifig row
for x in collection.find({"_id":1}):
    print(x)

# Get first 10 rows
for x in collection.find({}).sort({"_id":1}).limit(5):
    print(x)

## load dataframe

file_dataset = "./CSV/df_processed_polarity_and_satisfaction.csv"


df = pd.read_csv(file_dataset)
df.head()

to_upload = pd.DataFrame(df[["Id", "Polarity", "Customer_Satisfied"]])
to_upload.head()

for x in to_upload.head():
    print(x)
    

%%time
updates = []
for _, row in to_upload.iterrows():
    updates.append(UpdateOne({'Id': row.get('Id')}, {'$set': {'Polarity': row.get('Polarity')}}, upsert=True))

collection.bulk_write(updates)

%%time
updates = []
for _, row in to_upload.iterrows():
    updates.append(UpdateOne({'Id': row.get('Id')}, {'$set': {'Customer_Satisfied': row.get('Customer_Satisfied')}}, upsert=True))

collection.bulk_write(updates)

# Get first 10 rows
for x in collection.find({}).sort({"_id":1}).limit(10):
    print(x)
